+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## b1.11.0.4
+## b1.11.0.5
 - add Multidevicesupport for more then one endpoint. Documented in Readme.md
 - improved WLED Data Manager with endpoint-specific caching and automatic LED count detection
 - add SLEEP effect which can be activated after a configurable idle time. Activated with -SLE and -SLET arguments
 - add SLEOFF argument to configure the duration until WLED is turned off after entering sleep mode
 - removed mirror messages out of log. 
+- add playername specific IDLE effects with PIDE Argument 
 
 ## v1.10.4
 - Bugfix: Speed (sx), Intensity (ix) and Palette (pal) parameters are now correctly sent as integer values to WLED